refactor: 重构了后台账号管理页面的代码，将复杂组件拆分为多个简单组件
